+++ tests/__init__.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-"""Unit tests for the `yorm` package."""
+"""Integration tests for the `yorm` package."""
 
 import time
 import logging