+++ yorm/base/convertible.py
@@ -4,21 +4,27 @@ import abc
 
 from .. import common
 from . import MESSAGE
+from .converter import Converter
 
 
 log = common.logger(__name__)
 
 
-class Convertible(metaclass=abc.ABCMeta):
+class Convertible(Converter):  # pragma: no cover (abstract)
 
-    """Base class for attribute converters."""
+    """Base class for mutable attributes."""
 
-    @abc.abstractclassmethod
-    def to_value(cls, obj):
-        """Convert the loaded data back into the attribute's type."""
+    @classmethod
+    def to_value(cls, data):
+        value = cls.create_default()
+        value.update_value(data)
+        return value
+
+    @abc.abstractmethod
+    def update_value(self, data):  # pylint: disable=E0213,
+        """Update the attribute's value from loaded data."""
         raise NotImplementedError(MESSAGE)
 
-    @abc.abstractclassmethod
-    def to_data(cls, obj):
-        """Convert the attribute's value to data optimized for dumping."""
-        raise NotImplementedError(MESSAGE)
+    def format_data(self):
+        """Format the attribute to data optimized for dumping."""
+        return self.to_data(self)