+++ yorm/test/test_converters_extended.py
@@ -1,13 +1,36 @@
 #!/usr/bin/env python
-# pylint:disable=R0201
+# pylint:disable=R0201,R0901
 
 """Unit tests for the `extended` module."""
 
 import pytest
 
-from yorm.converters import Markdown, AttributeDictionary, SortedList
+from yorm.utilities import attr
+from yorm.converters.standard import Integer, String, Float
+from yorm.converters.extended import Markdown, AttributeDictionary, SortedList
 
-from .samples import *  # pylint: disable=W0401,W0614
+
+@attr(var1=Integer, var2=String)
+class SampleAttributeDictionary(AttributeDictionary):
+
+    """Sample dictionary container with initialization."""
+
+    def __init__(self, var1, var2, var3):
+        super().__init__()
+        self.var1 = var1
+        self.var2 = var2
+        self.var3 = var3
+
+
+@attr(all=Float)
+class SampleSortedList(SortedList):
+
+    """Sample sorted list container."""
+
+
+class UnknownSortedList(SortedList):
+
+    """Sample list container."""
 
 
 class TestMarkdown: